Enforced float policy
No floats are raised to double to and then casted back to float after
the computation. Replaced also the default M_PI with a float version
called ONE_PI (to maintain consistency with TWO_PI and FOUR_PI)

+++ src/renderer.cpp
@@ -196,7 +196,7 @@ int Renderer::render(Scene* s)
     //sizeof because MSVC cannot resolve strlen a compile time
     char endmsg[sizeof(MESSAGE_RENDERTIME)/sizeof(char)+MAX_TIME_FORMAT_LENGTH];
     char elapsed_formatted[16];
-    formatSeconds(duration_cast<seconds>(b-a).count(),elapsed_formatted);
+    formatSeconds((int)duration_cast<seconds>(b-a).count(),elapsed_formatted);
     sprintf(endmsg, MESSAGE_RENDERTIME,elapsed_formatted);
     Console.log(endmsg,NULL);
